server/requirements.txt
numpy
opencv-python-headless
Pillow
simple-lama-inpainting
